requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
openai==1.40.0
asyncpg==0.29.0
pyyaml==6.0.1
sqlglot==25.0.0
pydantic==2.8.0
httpx==0.27.0